evict the strings regex cache LRU instead of clearing it entirely
The cache dropped every compiled Pattern at once on reaching 1024 entries,
so a workload with many distinct regexes (dynamically built patterns, many
tenants sharing the JVM-wide cache) thrashed: hot, frequently-reused
patterns got wiped along with the rest, defeating the cache under exactly
the load it was meant for.

Moved ConcurrentLruCache out of MediaTypeUtils (where it was a private
nested class) into its own file so strings can reuse it instead of adding a
second bespoke bounded cache.
