+++ xtrasonnet/src/main/java/io/github/jam01/xtrasonnet/document/MediaTypeUtils.java
@@ -33,9 +33,11 @@ import org.jspecify.annotations.Nullable;
  * - Removed generateMultipartBoundary and related methods
  * - Refactored parameter parsing to method
  * - Only parse params if ; in media type
+ * - Moved the nested ConcurrentLruCache class to its own file, io.github.jam01.xtrasonnet.util.ConcurrentLruCache
  */
 
 import io.github.jam01.xtrasonnet.Utils;
+import io.github.jam01.xtrasonnet.util.ConcurrentLruCache;
 
 import java.nio.charset.UnsupportedCharsetException;
 import java.util.ArrayList;
@@ -47,11 +49,6 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentLinkedDeque;
-import java.util.concurrent.locks.ReadWriteLock;
-import java.util.concurrent.locks.ReentrantReadWriteLock;
-import java.util.function.Function;
 import java.util.stream.Collectors;
 
 /**
@@ -286,87 +283,6 @@ public abstract class MediaTypeUtils {
         Objects.requireNonNull(mimeTypes, "'mimeTypes' must not be null");
         if (mimeTypes.size() > 1) {
             mimeTypes.sort(SPECIFICITY_COMPARATOR);
-        }
-    }
-
-    /**
-     * Simple Least Recently Used cache, bounded by the maximum size given
-     * to the class constructor.
-     * <p>This implementation is backed by a {@code ConcurrentHashMap} for storing
-     * the cached values and a {@code ConcurrentLinkedQueue} for ordering the keys
-     * and choosing the least recently used key when the cache is at full capacity.
-     *
-     * @param <K> the type of the key used for caching
-     * @param <V> the type of the cached values
-     */
-    private static class ConcurrentLruCache<K, V> {
-
-        private final int maxSize;
-
-        private final ConcurrentLinkedDeque<K> queue = new ConcurrentLinkedDeque<>();
-
-        private final ConcurrentHashMap<K, V> cache = new ConcurrentHashMap<>();
-
-        private final ReadWriteLock lock;
-
-        private final Function<K, V> generator;
-
-        private volatile int size;
-
-        public ConcurrentLruCache(int maxSize, Function<K, V> generator) {
-            if (maxSize <= 0) {
-                throw new IllegalArgumentException("LRU max size should be positive");
-            }
-
-            Objects.requireNonNull(generator, "Generator function should not be null");
-            this.maxSize = maxSize;
-            this.generator = generator;
-            this.lock = new ReentrantReadWriteLock();
-        }
-
-        public V get(K key) {
-            V cached = this.cache.get(key);
-            if (cached != null) {
-                if (this.size < this.maxSize) {
-                    return cached;
-                }
-                this.lock.readLock().lock();
-                try {
-                    if (this.queue.removeLastOccurrence(key)) {
-                        this.queue.offer(key);
-                    }
-                    return cached;
-                } finally {
-                    this.lock.readLock().unlock();
-                }
-            }
-            this.lock.writeLock().lock();
-            try {
-                // Retrying in case of concurrent reads on the same key
-                cached = this.cache.get(key);
-                if (cached != null) {
-                    if (this.queue.removeLastOccurrence(key)) {
-                        this.queue.offer(key);
-                    }
-                    return cached;
-                }
-                // Generate value first, to prevent size inconsistency
-                V value = this.generator.apply(key);
-                int cacheSize = this.size;
-                if (cacheSize == this.maxSize) {
-                    K leastUsed = this.queue.poll();
-                    if (leastUsed != null) {
-                        this.cache.remove(leastUsed);
-                        cacheSize--;
-                    }
-                }
-                this.queue.offer(key);
-                this.cache.put(key, value);
-                this.size = cacheSize + 1;
-                return value;
-            } finally {
-                this.lock.writeLock().unlock();
-            }
         }
     }
 }
